# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for triggering database backups and managing them.",
        "title": "Database Backup API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/backup": {
            "post": {
                "description": "Queue a backup job for the specified database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backup"
                ],
                "summary": "Trigger a database backup",
                "parameters": [
                    {
                        "description": "Backup Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.BackupRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Backup started",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    },
                    "400": {
                        "description": "error: Bad request",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    }
                }
            }
        },
        "/backups": {
            "get": {
                "description": "Get a paginated list of all backups with optional filtering",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backup"
                ],
                "summary": "List all backups",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated status values (pending,generating,completed,failed)",
                        "name": "statuses",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search keyword (searches in database, host, type)",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "createdAt",
                        "description": "Field to order by",
                        "name": "orderBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "desc",
                        "description": "Order direction (asc/desc)",
                        "name": "orderDir",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start date for filtering (RFC3339 format)",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date for filtering (RFC3339 format)",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of backups",
                        "schema": {
                            "$ref": "#/definitions/model.BackupListResponse"
                        }
                    },
                    "500": {
                        "description": "error: Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    }
                }
            }
        },
        "/backups/stats": {
            "get": {
                "description": "Retrieve aggregated backup statistics by type and status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backup"
                ],
                "summary": "Get backup statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Start date for filtering (RFC3339 format)",
                        "name": "startDate",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End date for filtering (RFC3339 format)",
                        "name": "endDate",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Backup statistics",
                        "schema": {
                            "$ref": "#/definitions/database.BackupStats"
                        }
                    },
                    "500": {
                        "description": "error: Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    }
                }
            }
        },
        "/backups/{id}": {
            "get": {
                "description": "Retrieve a single backup by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backup"
                ],
                "summary": "Get a single backup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backup ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Backup details",
                        "schema": {
                            "$ref": "#/definitions/model.BackupMetadata"
                        }
                    },
                    "400": {
                        "description": "error: Bad request",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    },
                    "404": {
                        "description": "error: Backup not found",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    },
                    "500": {
                        "description": "error: Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a backup from both MongoDB and R2 storage",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backup"
                ],
                "summary": "Delete a backup",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backup ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Backup deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    },
                    "400": {
                        "description": "error: Bad request",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    },
                    "404": {
                        "description": "error: Backup not found",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    },
                    "500": {
                        "description": "error: Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    }
                }
            }
        },
        "/backups/{id}/download": {
            "get": {
                "description": "Generate a presigned URL to download a backup file from R2 storage",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "backup"
                ],
                "summary": "Download a backup file",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Backup ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Download URL",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "error: Bad request",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    },
                    "404": {
                        "description": "error: Backup not found",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    },
                    "500": {
                        "description": "error: Internal server error",
                        "schema": {
                            "$ref": "#/definitions/model.BackupResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "database.BackupStats": {
            "type": "object",
            "properties": {
                "byStatus": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "byType": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.BackupListResponse": {
            "type": "object",
            "properties": {
                "backups": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BackupMetadata"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "model.BackupMetadata": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "database": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "filePath": {
                    "type": "string"
                },
                "fileSize": {
                    "type": "integer"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "objectKey": {
                    "type": "string"
                },
                "status": {
                    "description": "pending, generating, completed, failed",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.BackupStatus"
                        }
                    ]
                },
                "timestamp": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.BackupRequest": {
            "type": "object",
            "properties": {
                "connectionUri": {
                    "type": "string",
                    "example": "mongodb://user:pass@host:port/db"
                },
                "database": {
                    "type": "string",
                    "example": "mydb"
                },
                "host": {
                    "type": "string",
                    "example": "localhost"
                },
                "password": {
                    "type": "string",
                    "example": "pass"
                },
                "port": {
                    "type": "string",
                    "example": "5432"
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.BackupType"
                        }
                    ],
                    "example": "postgre"
                },
                "username": {
                    "type": "string",
                    "example": "user"
                },
                "webhookUrl": {
                    "type": "string",
                    "example": "http://example.com/webhook"
                }
            }
        },
        "model.BackupResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "model.BackupStatus": {
            "type": "string",
            "enum": [
                "pending",
                "generating",
                "completed",
                "failed"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusGenerating",
                "StatusCompleted",
                "StatusFailed"
            ]
        },
        "model.BackupType": {
            "type": "string",
            "enum": [
                "postgre",
                "mysql",
                "mongo",
                "redis"
            ],
            "x-enum-varnames": [
                "Postgres",
                "MySQL",
                "Mongo",
                "Redis"
            ]
        },
...
    }
}